Add support for working on this project with nix
This uses a pinned git commit of nixpkgs-unstable to get GHC 8.10.4.
haskell-language-server is supported in the dev shell (`nix-shell` from
the repo root), although a wrapper script is necessary to run it in a
nix-shell.

An example of such a script is here:
```

nix-shell --run "haskell-language-server-wrapper $@"
```

You can also build the program with nix from any distro.

The upshot of this PR is that you can, if you have nix installed,
install from source or work on this software in one command, utilizing
binary caching to avoid having to build dependencies.